+++ asap_datasets/__init__.py
@@ -1,3 +1,3 @@
-from .data import *
+from .datatypes import *
 from .datasets import *
 from .schema import *
